+++ src/main/java/org/folio/rest/client/impl/PasswordResetActionClientImpl.java
@@ -6,7 +6,7 @@ import io.vertx.core.http.HttpClient;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.hc.core5.http.HttpStatus;
+import org.folio.HttpStatus;
 import org.folio.dbschema.ObjectMapperTool;
 import org.folio.rest.client.PasswordResetActionClient;
 import org.folio.rest.exception.OkapiModuleClientException;