+++ src/main/java/org/folio/service/password/UserPasswordServiceImpl.java
@@ -1,8 +1,6 @@
 package org.folio.service.password;
 
-import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
-import io.vertx.core.Handler;
 import io.vertx.core.MultiMap;
 import io.vertx.core.http.HttpClient;
 import io.vertx.core.http.HttpMethod;
@@ -37,42 +35,39 @@ public class UserPasswordServiceImpl implements UserPasswordService {
   }
 
   @Override
-  public UserPasswordService validateNewPassword(String userId, String newPassword, JsonObject okapiConnectionParams,
-                                                 Handler<AsyncResult<JsonObject>> asyncResultHandler) {
+  public Future<JsonObject> validateNewPassword(String userId, String newPassword, JsonObject okapiConnectionParams) {
     try {
       OkapiConnectionParams params = okapiConnectionParams.mapTo(OkapiConnectionParams.class);
       String url = params.getOkapiUrl() + VALIDATE_URL;
 
-      RestUtil.doRequest(httpClient, url, HttpMethod.POST, params.buildHeaders(), buildPasswordEntity(userId, newPassword))
-        .onComplete(h -> {
-          if (h.failed() || h.result().getCode() != 200) {
-            logger.error("Fail during sending request to validate password", h.cause());
-            asyncResultHandler.handle(Future.failedFuture(h.cause()));
-          } else {
-            JsonObject validateResult = h.result().getJson();
-            JsonArray messages = validateResult.getJsonArray("messages");
-            Errors errors = new Errors();
-            errors.setTotalRecords(0);
-            if (messages != null && !messages.isEmpty()) {
-              errors.setTotalRecords(messages.size());
-              List<Error> errorList = new ArrayList<>();
-              for (int i = 0; i < messages.size(); i++) {
-                Error error = new Error();
-                error.setMessage(messages.getString(i));
-                error.setCode(messages.getString(i));
-                errorList.add(error);
-              }
-              errors.setErrors(errorList);
-            }
-            asyncResultHandler.handle(Future.succeededFuture(JsonObject.mapFrom(errors)));
+      return RestUtil.doRequest(httpClient, url, HttpMethod.POST, params.buildHeaders(), buildPasswordEntity(userId, newPassword))
+        .compose(response -> {
+          if (response.getCode() != 200) {
+            logger.error("Fail during validating new password. Response code: {}, body: {}", response.getCode(), response.getBody());
+            return Future.failedFuture("Failed to validate new password. Status: %s, body: %s"
+              .formatted(response.getCode(), response.getBody()));
           }
+          JsonObject validateResult = response.getJson();
+          JsonArray messages = validateResult.getJsonArray("messages");
+          Errors errors = new Errors();
+          errors.setTotalRecords(0);
+          if (messages != null && !messages.isEmpty()) {
+            errors.setTotalRecords(messages.size());
+            List<Error> errorList = new ArrayList<>();
+            for (int i = 0; i < messages.size(); i++) {
+              Error error = new Error();
+              error.setMessage(messages.getString(i));
+              error.setCode(messages.getString(i));
+              errorList.add(error);
+            }
+            errors.setErrors(errorList);
+          }
+          return Future.succeededFuture(JsonObject.mapFrom(errors));
         });
     } catch (Exception e) {
       logger.error("Error during validating user's password", e);
-      asyncResultHandler.handle(Future.failedFuture(e));
-      return this;
+      return Future.failedFuture(e);
     }
-    return this;
   }
 
   private String buildPasswordEntity(String userId, String newPassword) {
@@ -83,11 +78,9 @@ public class UserPasswordServiceImpl implements UserPasswordService {
   }
 
   @Override
-  public UserPasswordService updateUserCredential(JsonObject newPasswordObject, JsonObject requestHeaders,
-                                                  Handler<AsyncResult<Integer>> asyncResultHandler) {
+  public Future<Integer> updateUserCredential(JsonObject newPasswordObject, JsonObject requestHeaders) {
     try {
       String url = requestHeaders.getString(BLUsersAPI.OKAPI_URL_HEADER) + UPDATE_URL;
-
       MultiMap headers = MultiMap.caseInsensitiveMultiMap();
       headers.addAll(requestHeaders
         .getMap()
@@ -95,25 +88,17 @@ public class UserPasswordServiceImpl implements UserPasswordService {
         .stream()
         .collect(Collectors.toMap(Map.Entry::getKey, e -> (String) e.getValue())));
 
-      RestUtil.doRequest(httpClient, url, HttpMethod.POST, headers, newPasswordObject.encode())
-        .onComplete(h -> {
-          if (h.failed()) {
-            logger.error("Fail during sending request to update user's credentials", h.cause());
-            asyncResultHandler.handle(Future.failedFuture(h.cause()));
-          } else {
-            int code = h.result().getCode();
-            if (code == 204 || code == 401) {
-              asyncResultHandler.handle(Future.succeededFuture(code));
-            } else {
-              asyncResultHandler.handle(Future.failedFuture(h.result().getBody()));
-            }
+      return RestUtil.doRequest(httpClient, url, HttpMethod.POST, headers, newPasswordObject.encode())
+        .compose(response -> {
+          int code = response.getCode();
+          if (code == 204 || code == 401) {
+            return Future.succeededFuture(code);
           }
+          return Future.failedFuture(response.getBody());
         });
     } catch (Exception e) {
       logger.error("Error during updating user's credentials", e);
-      asyncResultHandler.handle(Future.failedFuture(e));
-      return this;
+      return Future.failedFuture(e);
     }
-    return this;
   }
 }